--- a/eq-author-api/middleware/export.js
+++ b/eq-author-api/middleware/export.js
@@ -1,0 +1,8 @@
+const { getQuestionnaire } = require("../utils/datastore");
+
+module.exports = async (req, res) => {
+  const questionnaireId = req.params.questionnaireId;
+  const questionnaire = await getQuestionnaire(questionnaireId);
+
+  res.status(200).json(questionnaire);
+};