--- a/eq-author-api/middleware/import.js
+++ b/eq-author-api/middleware/import.js
@@ -1,0 +1,12 @@
+const uuid = require("uuid");
+const { createQuestionnaire } = require("../utils/datastore");
+
+module.exports = async (req, res) => {
+  const input = req.body;
+  input.id = uuid.v4();
+  input.createdAt = new Date();
+  const savedQuestionnaire = await createQuestionnaire(input);
+  res.json({
+    id: savedQuestionnaire.id,
+  });
+};