--- a/eq-author-api/middleware/import.js
+++ b/eq-author-api/middleware/import.js
@@ -2,7 +2,6 @@ const uuid = require("uuid");
 const { createQuestionnaire } = require("../utils/datastore");
 
 module.exports = async (req, res) => {
-  console.log(req.body);
   const input = req.body;
   input.id = uuid.v4();
   input.createdAt = new Date();