--- a/eq-author/cypress/fixtures/seed/test-routing-and.json
+++ b/eq-author/cypress/fixtures/seed/test-routing-and.json
@@ -1,0 +1,176 @@
+{
+  "summary": false,
+  "metadata": [
+    {
+      "id": "e7ff5c9b-b8da-435a-bc03-768a65a7b644",
+      "key": "ru_name",
+      "alias": "Ru Name",
+      "type": "Text",
+      "textValue": "ESSENTIAL ENTERPRISE LTD."
+    },
+    {
+      "id": "dd5fd4dd-bc9f-4d04-a639-2b0b90acf5d8",
+      "key": "trad_as",
+      "alias": "Trad As",
+      "type": "Text",
+      "textValue": "ESSENTIAL ENTERPRISE LTD."
+    },
+    {
+      "id": "f7ceb5e9-6e99-494e-a887-9592c834b219",
+      "key": "period_id",
+      "alias": "Period Id",
+      "type": "Text",
+      "textValue": "201605"
+    },
+    {
+      "id": "b86a5ad0-17c0-4722-9b0c-940d0f39f2cf",
+      "key": "ref_p_start_date",
+      "alias": "Start Date",
+      "type": "Date",
+      "dateValue": "2016-05-01"
+    },
+    {
+      "id": "b0d49080-dfee-43f1-99b7-3a9bd5eb0eb4",
+      "key": "ref_p_end_date",
+      "alias": "End Date",
+      "type": "Date",
+      "dateValue": "2016-06-12"
+    },
+    {
+      "id": "54015482-0f20-4176-ae63-6362b4c22efb",
+      "key": "employmentDate",
+      "alias": "Employment Date",
+      "type": "Date",
+      "dateValue": "2016-06-10"
+    }
+  ],
+  "legalBasis": "StatisticsOfTradeAct",
+  "title": "Test routing AND",
+  "type": "Business",
+  "version": 4,
+  "sections": [
+    {
+      "id": "dc6a990e-ec77-4595-91a5-f6e1d9b604ad",
+      "title": "<p>Section 1</p>",
+      "introductionEnabled": false,
+      "pages": [
+        {
+          "id": "7a3ec23c-e984-4ea5-8a9a-cc29859e89b2",
+          "pageType": "QuestionPage",
+          "title": "<p>Question 1</p>",
+          "description": "",
+          "descriptionEnabled": false,
+          "guidanceEnabled": false,
+          "definitionEnabled": false,
+          "additionalInfoEnabled": false,
+          "answers": [
+            {
+              "id": "0a6bc043-4e38-46f8-9dbb-b60b6cc28221",
+              "description": "",
+              "guidance": "",
+              "label": "Answer 1",
+              "qCode": "",
+              "type": "Radio",
+              "questionPageId": "7a3ec23c-e984-4ea5-8a9a-cc29859e89b2",
+              "properties": {
+                "required": false
+              },
+              "validation": {},
+              "options": [
+                {
+                  "id": "117a30ef-214d-470b-993c-09f08ff6357c",
+                  "label": "A",
+                  "description": null,
+                  "additionalAnswer": null
+                },
+                {
+                  "id": "9a987637-3a88-4274-8db2-cdc260b3f42a",
+                  "label": "B",
+                  "description": null,
+                  "additionalAnswer": null
+                },
+                {
+                  "id": "cddb080b-ee50-4ee3-90f9-e7945e67f165",
+                  "label": "C",
+                  "description": null,
+                  "additionalAnswer": null
+                }
+              ],
+              "secondaryLabel": null
+            }
+          ],
+          "routing": null,
+          "alias": null,
+          "guidance": null,
+          "definitionLabel": null,
+          "definitionContent": null,
+          "additionalInfoLabel": null,
+          "additionalInfoContent": null
+        },
+        {
+          "id": "4bcc06e1-8090-4275-9cea-de6432c800db",
+          "pageType": "QuestionPage",
+          "title": "<p>Question 2</p>",
+          "description": "",
+          "descriptionEnabled": false,
+          "guidanceEnabled": false,
+          "definitionEnabled": false,
+          "additionalInfoEnabled": false,
+          "answers": [
+            {
+              "id": "5eef5520-f871-43ca-a8f9-2b65a6a3f727",
+              "description": "",
+              "guidance": "",
+              "label": "Answer 2",
+              "qCode": "",
+              "type": "Radio",
+              "questionPageId": "4bcc06e1-8090-4275-9cea-de6432c800db",
+              "properties": {
+                "required": false
+              },
+              "validation": {},
+              "options": [
+                {
+                  "id": "aea46eec-0b8a-4e0d-89d8-bc1905612d2d",
+                  "label": "D",
+                  "description": null,
+                  "additionalAnswer": null
+                },
+                {
+                  "id": "501f9b81-b9df-416e-a51a-211c255b135b",
+                  "label": "E",
+                  "description": null,
+                  "additionalAnswer": null
+                },
+                {
+                  "id": "de900da7-e226-40ce-9989-a63d1c784b53",
+                  "label": "F",
+                  "description": null,
+                  "additionalAnswer": null
+                }
+              ],
+              "secondaryLabel": null
+            }
+          ],
+          "routing": null,
+          "alias": null,
+          "sectionId": "dc6a990e-ec77-4595-91a5-f6e1d9b604ad",
+          "guidance": null,
+          "definitionLabel": null,
+          "definitionContent": null,
+          "additionalInfoLabel": null,
+          "additionalInfoContent": null
+        }
+      ],
+      "alias": "",
+      "introductionTitle": null,
+      "introductionContent": null
+    }
+  ],
+  "createdAt": "2019-04-23T11:34:28.931Z",
+  "navigation": true,
+  "createdBy": "Cypress",
+  "theme": "default",
+  "id": "145d591b-a39a-4385-a584-f3d9092e8e6c",
+  "updatedAt": "2019-04-23T11:34:36.137Z"
+}