### What is the context of this PR?
This is a spike to test the performance improvement of seeding cypress tests it:
1. Creates an import endpoint
2. Creates an export endpoint
3. Migrates the `routing_spec` to use it.

### Performance
master
```
routing: 2 minutes, 15 seconds
validation: 4 minutes, 46 seconds 
```
this branch
```
routing: 1 minute, 4 seconds
validation: 2 minutes, 38 seconds 
```

### How to review 
Don't yet.